[iOS][updates] Migrate tests to Swift Testing (1/3)
